Restructure useStates and add map logic to render each catch
